+++ src/main/java/com/sedikev/domain/model/LoteDomain.java
@@ -2,6 +2,7 @@ package com.sedikev.domain.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -12,6 +13,6 @@ public class LoteDomain {
     private Long id;
     private UsuarioDomain usuarioDomain;
     private Integer contramarca;
-    private Float precio_kilo;
+    private BigDecimal precio_kilo;
     private Date fecha;
 }
