+++ src/main/java/com/sedikev/infrastructure/adapter/entity/GastoEntity.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -29,7 +30,7 @@ public class GastoEntity {
     private UsuarioEntity usuarioEntity;
 
     @Column(name = "cantidad")
-    private Float cantidad;
+    private BigDecimal cantidad;
 
     @Column(name = "descripcion")
     @Lob
