+++ src/main/java/com/sedikev/infrastructure/adapter/entity/PagoEntity.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -32,7 +33,7 @@ public class PagoEntity {
     private String tipo_pago;
 
     @Column(name = "cantidad")
-    private Float cantidad;
+    private BigDecimal cantidad;
 
     @Column(name = "descripcion")
     @Lob
